finish catalog logic. reorganize files.
