--- src/components/Catalog/types.ts
@@ -1,124 +1,0 @@
-import type { ReactiveControllerHost } from "lit";
-import type { Variable } from "./browse-variables.types.js";
-
-export type SearchOptions = {
-  signal?: AbortSignal;
-  bearerToken?: string;
-};
-
-export interface MetadataCatalogInterface {
-  getCollectionCitation(
-    collectionEntryId: string,
-    options?: SearchOptions
-  ): Promise<CmrCollectionCitationItem>;
-
-  searchCmr(
-    keyword: string,
-    type: "collection" | "variable" | "all",
-    options?: SearchOptions
-  ): Promise<Array<CmrSearchResult>>;
-}
-
-export type CmrCollectionCitationsResponse = {
-  collections: {
-    items: Array<CmrCollectionCitationItem>;
-  };
-};
-
-export type CmrCollectionCitationItem = {
-  doi: {
-    doi: string;
-  };
-  collectionCitations: Array<CmrCollectionCitation>;
-};
-
-export type CmrCollectionCitation = {
-  creator: string;
-  editor: string;
-  dataPresentationForm: string;
-  onlineResource: {
-    linkage: string;
-  };
-  publisher: string;
-  title: string;
-  seriesName: string;
-  releaseDate: string;
-  version: string;
-  releasePlace: string;
-};
-
-export type CmrSearchResult = {
-  type: "collection" | "variable";
-  collectionConceptId: string;
-  collectionEntryId: string;
-  summary: string;
-  conceptId: string;
-  entryId: string;
-  provider: string;
-  title: string;
-};
-
-export type CmrSearchResultsResponse = {
-  collections?: {
-    items: Array<{
-      conceptId: string;
-      nativeId: string;
-      provider: string;
-      title: string;
-    }>;
-  };
-  variables?: {
-    items: Array<{
-      conceptId: string;
-      name: string;
-      providerId: string;
-      longName: string;
-      collections: {
-        items: Array<{
-          conceptId: string;
-          nativeId: string;
-          title: string;
-        }>;
-      };
-    }>;
-  };
-};
-
-export interface VariableCatalogInterface {
-  /**
-   * Fetches the list of search keywords
-   * @returns Promise containing the list of search keywords
-   */
-
-  getSearchKeywords(): Promise<SearchKeywordsResponse>;
-}
-
-export type SearchKeywordsResponse = {
-  id: string;
-};
-
-export type GiovanniFacetValue = {
-  name: string;
-  count: number;
-};
-
-export type GiovanniFacet = {
-  category: string;
-  values: GiovanniFacetValue[];
-};
-
-export type GetVariablesResponse = {
-  count: number;
-  total: number;
-  variables: Variable[];
-  facets: GiovanniFacet[];
-};
-
-export type HostWithMaybeProperties = ReactiveControllerHost & {
-  variableEntryId?: string;
-  collection?: string;
-  variable?: string;
-  startDate?: string;
-  endDate?: string;
-  catalogVariable?: Variable;
-};
