--- src/store/NetworkContext.tsx
@@ -1,42 +1,0 @@
-import React, { createContext, useEffect, useState, useContext } from "react";
-import { Network } from "@capacitor/network";
-
-interface NetworkContextType {
-  isOnline: boolean;
-}
-
-const NetworkContext = createContext<NetworkContextType>({ isOnline: true });
-
-export const NetworkProvider: React.FC<{ children: React.ReactNode }> = ({
-  children,
-}) => {
-  const [isOnline, setIsOnline] = useState(true);
-
-  useEffect(() => {
-    // Get initial network state
-    Network.getStatus().then((status) => setIsOnline(status.connected));
-
-    // Listen for network status changes
-    const listener = Network.addListener("networkStatusChange", (status) => {
-      setIsOnline(status.connected);
-    });
-
-    return () => {
-      listener.then((l) => l.remove());
-    };
-  }, []);
-
-  return (
-    <NetworkContext.Provider value={{ isOnline }}>
-      {children}
-    </NetworkContext.Provider>
-  );
-};
-
-export const useNetwork = () => {
-  const context = useContext(NetworkContext);
-  if (context === undefined) {
-    throw new Error("useNetwork must be used within a NetworkProvider");
-  }
-  return context;
-};
